Teach Mod how to install and update itself
